Ship fixtures
This allow job-server to reuse them.

Also exposes the core API via __init__.py, for cleanliness
